Change const to constexpr and unsigned int to uint_(8/16)t
This change allow speed up program in run-time and use less memory.

+++ rooms-hierarchy/room.hpp
@@ -23,7 +23,7 @@ protected:
 
   std::vector<Team> teams_;
   std::string referee_;
-  unsigned int classroom_;
+  uint16_t classroom_;
 };
 
 std::ostream &operator<<(std::ostream &out, const Room &room);
